+++ services/ui-src/src/measures/2022/shared/globalValidations/index.ts
@@ -17,7 +17,7 @@ export * from "shared/globalValidations/validateOneQualDenomHigherThanOtherDenom
 export * from "./validateOneQualRateHigherThanOtherQual";
 export * from "shared/globalValidations/validateReasonForNotReporting";
 export * from "shared/globalValidations/validateRequiredRadioButtonForCombinedRates";
-export * from "./validateTotals";
+export * from "shared/globalValidations/validateTotals";
 export * from "shared/globalValidations/validateYearFormat";
 
 // PCR-XX Specific Validations
