+++ services/ui-src/src/measures/2023/shared/util/validationsMock.tsx
@@ -17,7 +17,7 @@ import * as validateOneQualDenomHigherThanOtherDenomOMS from "shared/globalValid
 import * as validateOneQualRateHigherThanOtherQual from "measures/2023/shared/globalValidations/validateOneQualRateHigherThanOtherQual";
 import * as validateReasonForNotReporting from "shared/globalValidations/validateReasonForNotReporting";
 import * as validateRequiredRadioButtonForCombinedRates from "shared/globalValidations/validateRequiredRadioButtonForCombinedRates";
-import * as validateTotals from "measures/2023/shared/globalValidations/validateTotals";
+import * as validateTotals from "shared/globalValidations/validateTotals";
 import * as PCRatLeastOneRateComplete from "shared/globalValidations/PCRValidations/PCRatLeastOneRateComplete";
 import * as PCRnoNonZeroNumOrDenom from "shared/globalValidations/PCRValidations/PCRnoNonZeroNumOrDenom";
 import * as ComplexAtLeastOneRateComplete from "shared/globalValidations/ComplexValidations/ComplexAtLeastOneRateComplete";
