+++ services/ui-src/src/measures/2024/shared/globalValidations/index.ts
@@ -23,7 +23,7 @@ export * from "shared/globalValidations/validateOneQualDenomHigherThanOtherDenom
 export * from "./validateOneQualRateHigherThanOtherQual";
 export * from "shared/globalValidations/validateReasonForNotReporting";
 export * from "shared/globalValidations/validateRequiredRadioButtonForCombinedRates";
-export * from "./validateTotals";
+export * from "shared/globalValidations/validateTotals";
 export * from "shared/globalValidations/validateYearFormat";
 export * from "shared/globalValidations/validateOPMRates";
 export * from "shared/globalValidations/validateHedisYear";
