+++ services/ui-src/src/shared/globalValidations/validateTotals/index.test.tsx
